+++ jgclark.Dashboard/CHANGELOG.md
@@ -1,8 +1,11 @@
 # What's changed in 🎛 Dashboard plugin?
-For more details see the [plugin's README](https://github.com/NotePlan/plugins/tree/main/jgclark.Dashboard/).
+For more details see the [plugin's documentation](https://github.com/NotePlan/plugins/tree/main/jgclark.Dashboard/).
 
-## [0.7.5] - 2024-01-03 (unreleased)
-- added action button '≯' to 'unschedule' a task. Note: only relevant on tasks in Project notes.
+## [0.8.0] - 2024-01-06
+- added action button '≯' to 'unschedule' a task. (This is only relevant on tasks referenced/scheduled to calendar notes.)
+- added action button '✓then' to complete an overdue task, but marks it `@done(...)` at the original time, not today.
+- added action buttons 'skip ...' to items in the Project section. (This does the equivalent of "/skip project review", skipping the review of that note for the period shown.)
+- added new 'Theme to use for Dashboard' setting for @anton.skliar. If this is set to a valid Theme name from among those you have installed, this Theme will be used instead of your current Theme.
 
 ## [0.7.4] - 2023-12-26
 ### Added